+++ docs/source/conf.py
@@ -167,7 +167,7 @@ author = u'Jon Griffiths'
 # built documents.
 #
 # The short X.Y version.
-version = u'1.0.0'
+version = u'1.1.0'
 # The full version, including alpha/beta/rc tags.
 release = version