+++ include/wally_core.h
@@ -30,9 +30,9 @@ extern "C" {
 
 /** Library version */
 #define WALLY_MAJOR_VER 1
-#define WALLY_MINOR_VER 0
+#define WALLY_MINOR_VER 1
 #define WALLY_PATCH_VER 0
-#define WALLY_BUILD_VER 0x10000
+#define WALLY_BUILD_VER 0x10100
 
 /**
  * Initialize wally.